Only analyse TypeSpecs once. Fixes #29.
Previously, the type specifiers of declarations with multiple
declarators were analyzed multiple times. However, when declaring a
variable with the same name as a type, this causes problems.

Consider:

  foo *foo, *bar;

The analysis of "foo *foo" results in a symbol table where the innermost
definition of "foo" refers to an object, not a typedef name.
Therefore, when analyzing "foo *bar", the type checker tries to look up
"foo" and gets an object declaration, which is a type error.

Code like the above example does actually occur, and is valid C. The
seemingly equivalent:

  foo *foo;
  foo *bar;

is not valid C, and is rejected as it should be. The second line is seen
as an expression statement, multiplying foo and bar, and fails because
bar is not in scope.

darcs-hash:5ad2849ca02b3c699328ef606e5976572988fe85