Further progress on bug #29
Ignore-this: d2b78ed617dbd4c3d70df5a92e1e9e4e

It's now possible to correctly type check this:

    typedef int ax25_dev;
    int f() {
        ax25_dev *ax25_dev, *ax25_dev_old;
    }

and this:

    typedef int ax25_dev;
    int f() {
        typeof(ax25_dev) *ax25_dev, *ax25_dev_old;
    }

but not this:

    typedef int ax25_dev;
    int f() {
        typeof(ax25_dev) *ax25_dev, *ax25_dev_old;
        typeof(ax25_dev_old) foo;
    }

The second line resolves to ax25_dev, which now is an object, not a
type. Is this valid ANSI C? GCC accepts it.

darcs-hash:0634a812e9db8e4ea60a61eea4d9b18fc56d7d63