Fix type checking bug for function pointer parameters.
The previous patch to add support for transparent unions broke type
checking of paramters to functions passed via function pointers. If a
function was defined to take a parameter that was a typedef name, and
you passed in a function in which the matching parameter was the raw
type referred to by that typedef, the two would not match because the
type checker was failing to dereference typedef names in this one case.

darcs-hash:cee550a1d715b804ed6ebbb8c6e46241ea7d2ea8